chore: cleanup and add new sharing touchpoint
