--- app/src/components/features/sharedLists/CreateSharedListModal/index.css
@@ -1,3 +1,0 @@
-.restricted-users-input.react-multi-email > input {
-  width: 100% !important;
-}
